Move new constraints to further constraints
Deleted data utils
Fixed dependency or resources and processes in further constraints
